requirements.txt
flask>=3.0.0
flask-bcrypt>=1.0.1
flask-limiter>=3.5.0
flask-wtf>=1.2.1
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.32.0
